pyproject.toml
[project]
name = "agile-monkeys-assestment"
version = "1.0.0"
description = "The Agile Monkeys Assestment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.6",
]

[tool.ruff]
# Set the maximum line length to 88 (Black's default)
line-length = 88

# Assume Python 3.12+ (adjust based on your version)
target-version = "py312"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".pytest_cache",
]

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort (import sorting)
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
]

# Ignore specific rules if needed
ignore = [
    "E501",  # line too long (handled by formatter)
]

# Allow fix for all enabled rules (when `--fix` is provided)
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Use `\n` line endings for all files
line-ending = "auto"
